armour -> armor, LivingSet -> LivingData, moved UpgradeHolderBase to API
