+++ src/main/java/wayoftime/bloodmagic/api/item/UpgradeHolderBase.java
