+++ src/main/java/wayoftime/bloodmagic/common/datamap/LivingArmorData.java
