+++ src/main/java/wayoftime/bloodmagic/common/item/LivingArmorItem.java
